Split relations and relation-fk
